+++ src/content/index.tsx
@@ -1,3 +1,4 @@
+import { Hotkey, Hotkeys, HotkeysTarget } from "@blueprintjs/core";
 import { l } from "@r2c/extension/analytics";
 import { ApiFetch, getAnalyticsParams } from "@r2c/extension/api/fetch";
 import {
@@ -54,6 +55,7 @@ export const ExtensionContext = React.createContext<ContentHostState>(
   DEFAULT_STATE
 );
 
+@HotkeysTarget
 export default class ContentHost extends React.Component<{}, ContentHostState> {
   public state: ContentHostState = DEFAULT_STATE;
 
@@ -185,6 +187,23 @@ export default class ContentHost extends React.Component<{}, ContentHostState> {
       </div>
     );
   }
+
+  public renderHotkeys() {
+    return (
+      <Hotkeys>
+        <Hotkey
+          combo="esc"
+          label="Close Preflight side-panel"
+          onKeyDown={this.handleCloseHotKey}
+        />
+        ;
+      </Hotkeys>
+    );
+  }
+
+  private handleCloseHotKey = () => {
+    this.setState({ twistTab: undefined });
+  };
 
   private handleChecklistItemClick = (itemType: PreflightChecklistItemType) => (
     e: React.MouseEvent<HTMLElement>